+++ src/i18n_check/check/repeat_values.py
@@ -44,6 +44,7 @@ def get_repeat_value_counts(i18n_src_dict: Dict[str, str]) -> Dict[str, int]:
     Dict[str, int]
         A dictionary with values that appear more than once, mapped to their count.
     """
+    # Note: The following automatically removes repeat keys from i18n_src_dict.
     all_json_values = [
         lower_and_remove_punctuation(text=v)
         for v in list(i18n_src_dict.values())
@@ -53,11 +54,11 @@ def get_repeat_value_counts(i18n_src_dict: Dict[str, str]) -> Dict[str, int]:
     return {k: v for k, v in dict(Counter(all_json_values)).items() if v > 1}
 
 
-def analyze_and_suggest_keys(
+def analyze_and_generate_repeat_value_report(
     i18n_src_dict: Dict[str, str], json_repeat_value_counts: Dict[str, int]
 ) -> Dict[str, int]:
     """
-    Analyze repeated values and suggest solutions for key modification or removal.
+    Analyze repeated values and generates a report of repeat values with changes that should be made.
 
     Parameters
     ----------
@@ -69,9 +70,11 @@ def analyze_and_suggest_keys(
 
     Returns
     -------
-    Dict[str, int]
-        The updated dictionary of repeat value counts after suggested changes.
+    Dict[str, int], str
+        The updated dictionary of repeat value counts after suggested changes and a report to be added to the error.
     """
+    repeat_value_error_report = ""
+
     keys_to_remove = []
     for repeat_value in json_repeat_value_counts:
         i18n_keys = [
@@ -83,9 +86,11 @@ def analyze_and_suggest_keys(
 
         # Needed as we're removing keys that are set to lowercase above.
         if len(i18n_keys) > 1:
-            print(f"\nRepeat value: '{repeat_value}'")
-            print(f"Number of instances: : {json_repeat_value_counts[repeat_value]}")
-            print(f"Keys: {', '.join(i18n_keys)}")
+            repeat_value_error_report += (
+                f"\nRepeat value: '{repeat_value}'"
+                f"\nNumber of instances: : {json_repeat_value_counts[repeat_value]}"
+                f"\nKeys: {', '.join(i18n_keys)}"
+            )
 
             common_prefix = ""
             min_key_length = min(len(k) for k in i18n_keys)
@@ -105,10 +110,14 @@ def analyze_and_suggest_keys(
             if common_prefix := ".".join(common_prefix.split(".")[:-1]).replace(
                 "._global", ""
             ):
-                print(f"Suggested new key: {common_prefix}._global.IDENTIFIER_KEY")
+                repeat_value_error_report += (
+                    f"\nSuggested new key: {common_prefix}._global.CONTENT_REFERENCE"
+                )
 
             else:
-                print("Suggested new key: i18n._global.IDENTIFIER_KEY")
+                repeat_value_error_report += (
+                    "\nSuggested new key: i18n._global.CONTENT_REFERENCE"
+                )
 
         else:
             # Remove the key if the repeat is caused by a lowercase word.
@@ -117,13 +126,15 @@ def analyze_and_suggest_keys(
     for k in keys_to_remove:
         json_repeat_value_counts.pop(k, None)
 
-    return json_repeat_value_counts
+    return json_repeat_value_counts, repeat_value_error_report
 
 
 # MARK: Error Outputs
 
 
-def validate_repeat_values(json_repeat_value_counts: Dict[str, int]) -> None:
+def validate_repeat_values(
+    json_repeat_value_counts: Dict[str, int], repeat_value_error_report: str
+) -> None:
     """
     Check and report if there are repeat translation values.
 
@@ -131,6 +142,9 @@ def validate_repeat_values(json_repeat_value_counts: Dict[str, int]) -> None:
     ----------
     json_repeat_value_counts : Dict[str, int]
         A dictionary with repeat i18n values and their counts.
+
+    repeat_value_error_report : str
+        An error report including repeat values and changes that should be made.
 
     Returns
     -------
@@ -143,17 +157,17 @@ def validate_repeat_values(json_repeat_value_counts: Dict[str, int]) -> None:
         The system exits with 1 and prints error details if repeat values are found.
     """
     if json_repeat_value_counts:
+        is_or_are = "is"
+        it_or_them = "it"
         value_or_values = "value"
-        if len(json_repeat_value_counts) == 1:
-            value_to_be = "value is"
-
-        else:
+        if len(json_repeat_value_counts) > 1:
+            is_or_are = "are"
+            it_or_them = "them"
             value_or_values = "values"
-            value_to_be = "values are"
 
         error_message = "\n[red]"
-        error_message += f"❌ repeat_values error: {len(json_repeat_value_counts)} repeat i18n {value_to_be} present in the i18n source file. Please combine the {value_or_values} below into one key:\n\n"
-        error_message += "\n".join(json_repeat_value_counts.keys())
+        error_message += f"❌ repeat_values error: There {is_or_are} {len(json_repeat_value_counts)} repeat i18n {value_or_values} present in the i18n source file. Please follow the directions below to combine {it_or_them} into one key:\n"
+        error_message += repeat_value_error_report
         error_message += "[/red]"
 
         rprint(error_message)
@@ -171,7 +185,13 @@ def validate_repeat_values(json_repeat_value_counts: Dict[str, int]) -> None:
 
 if __name__ == "__main__":
     json_repeat_value_counts = get_repeat_value_counts(i18n_src_dict)
-    validate_repeat_values(json_repeat_value_counts=json_repeat_value_counts)
-    analyze_and_suggest_keys(
-        i18n_src_dict=i18n_src_dict, json_repeat_value_counts=json_repeat_value_counts
+    json_repeat_value_counts, repeat_value_error_report = (
+        analyze_and_generate_repeat_value_report(
+            i18n_src_dict=i18n_src_dict,
+            json_repeat_value_counts=json_repeat_value_counts,
+        )
+    )
+    validate_repeat_values(
+        json_repeat_value_counts=json_repeat_value_counts,
+        repeat_value_error_report=repeat_value_error_report,
     )