+++ src/i18n_check/check/unused_keys.py
@@ -106,7 +106,7 @@ def print_unused_keys(unused_keys: List[str]) -> None:
             "[red]\n❌ unused_keys error: There "
             + f"{to_be} {len(unused_keys)} i18n {key_to_be} unused. Please remove or assign the following {key_or_keys}:"
             + "\n\n"
-            + ", ".join(unused_keys)
+            + "\n".join(unused_keys)
             + "[/red]"
         )
         rprint(error_message)