+++ src/i18n_check/test_frontends/all_checks_fail/test_file.ts
@@ -1,8 +1,10 @@
 // This file contains the following i18n keys:
 //   - "i18n._global.hello_global"
-//   - "i18n._global.hello_global_repeat"
-//   - "i18n.wrong_identifier_path.incorrectly_named_key"
+//   - "i18n._global.hello_global_repeat_value"
+//   - "i18n._global.repeat_key"
 //   - i18n.test_file.hello_test_file  // no quotes or backticks
 //   - "i18n.test_file.incorrectly-formatted-key"
 //   - "i18n.test_file.nested_example"
 //   - "i18n.test_file.not_in_i18n_source_file"
+//   - "i18n.test_file.repeat_key_lower"  // lower cancels the repeat value check
+//   - "i18n.wrong_identifier_path.content_reference"